# time_library api gateway/src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/event": {
      "get": {
        "tags": [
          "Search Text"
        ],
        "summary": "Get events by id or name",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id of the event",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "name of the event",
            "required": false,
            "type": "string"
          }
        ]
      },
      "post": {
        "tags": [
          "Search Text"
        ],
        "summary": "Creates a new event",
        "conumes": "application/json",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "Event",
            "in": "body",
            "description": "The event to create",
            "schema": {"$ref":"#/definitions/Event"}
          }
        ]
      },
      "put": {
        "tags": [
          "Search Text"
        ],
        "summary": "Updates an event",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id of the event",
            "required": true,
            "type": "string"
          },
          {
            "name": "Event",
            "in": "body",
            "description": "The event to create",
            "schema": {"$ref":"#/definitions/Event"}
          }
        ]
      }
    },
    "/timeline": {
      "get": {
        "tags": [
          "Search Text"
        ],
        "summary": "Retrieve a list of timelines",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": []
      }
    },
    "/login": {
      "get": {
        "tags": [
          "Search Text"
        ],
        "summary": "Login via a provider (Google or Facebook OAuth)",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": []
      }
    }
  },
  "definitions": {
    "Event": {
      "type": "object",
      "required": [
        "name",
        "description"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    }
  }
}